+++ deploy-site.sh
@@ -1,2 +1,4 @@
+cd examples/docs 
+
 pnpm run site
 pnpm exec gh-pages -d dist-site
