+++ examples/docs/tailwind.config.js
@@ -1,6 +1,6 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   darkMode: ['class'],
-  content: ['./src/**/*.{js,ts,jsx,tsx,mdx}', './index.html', './main.tsx'],
+  content: ['./src/**/*.{js,ts,jsx,tsx,mdx}', './index.html'],
   theme: {}
 };
