+++ package.json
@@ -11,7 +11,6 @@
   "scripts": {
     "build": "tsc && pnpm run build:css",
     "build:css": "tailwindcss -i ./style/index.css -o ./dist/style.css",
-    "site": "tsc && vite build -c ./vite.config.site.ts",
     "lint": "eslint src --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "format": "pnpm run format:source && pnpm run format:other",
     "format:other": "prettier ./**/*.{md,json,html} --write",
@@ -77,8 +76,6 @@
     "eslint": "^8.38.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.3.4",
-    "file-saver": "^2.0.5",
-    "gh-pages": "^5.0.0",
     "postcss": "8.4.24",
     "prettier": "^3.0.0",
     "tailwindcss": "3.3.2",
